+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In progress
 
 - `logarithmic` is now `true` by default
+- Add `pattern` and `flash` method with `SpeedSetting` parameter
 - Added example `FadeComparison`
 
 ## v2.2.0 (2021-12-17)